Fix indentation on yaml snippets
The snippets of yaml in the enumerated list are indented by 3 spaces. Something in the rendering absorbs one of these spaces, which causes indentation of the first indent to be 1 space, whereas all subsequent indents are 2 spaces.

I've removed 1 space from the indentation of the entire `prism` block, which I'm hoping will fix the issue. I've deliberately not removed *all* indentation of the `prism` block because that makes it less clear (for humans, I don't know how the markdown rendering would respond) that these are part of each element in the enumerated list.

I've also changed the indentation of the `prism` block for PHP snippet on this page, for consistency with the updated yaml snippets.